requirements.txt
adium-theme-ubuntu==0.3.4
aniso8601==1.2.0
ansible==2.0.0.2
bcrypt==3.1.0
beautifulsoup4==4.5.3
bigsuds==1.0.4
blinker==1.4
cffi==1.7.0
chardet==2.3.0
click==6.6
contextlib2==0.5.4
cryptography==1.7.1
dnspython==1.15.0
ecdsa==0.13
enum34==1.1.6
Flask==0.11.1
Flask-Cors==3.0.0
flask-ldap3-login==0.9.12
Flask-Login==0.4.0
Flask-RESTful==0.3.5
flask-restful-swagger-2==0.30
Flask-WTF==0.14.2
funcsigs==1.0.2
html5lib==0.999
httplib2==0.9.1
idna==2.2
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.8
jsonify==0.5
ldap3==2.2.1
lxml==3.7.2
MarkupSafe==0.23
mock==2.0.0
ncclient==0.5.3
neo4j-driver==1.0.2
neo4jrestclient==2.1.1
netaddr==0.7.19
netmiko==1.2.5
paramiko==2.1.1
parse==1.6.6
passlib==1.6.5
pbr==1.10.0
py2neo==3.1.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
python-dateutil==2.6.0
pytz==2016.10
pyvmomi==6.5
PyYAML==3.12
requests==2.13.0
scour==0.32
scp==0.10.2
six==1.10.0
suds==0.4
unity-lens-photos==1.0
vcrpy==1.10.3
Werkzeug==0.11.10
wrapt==1.10.8
WTForms==2.1
XlsxWriter==0.9.3
xmltodict==0.10.2
